requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiodns==4.0.0
    # via aiohttp
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   aiohttp-retry
    #   runpod
aiohttp-retry==2.9.1
    # via runpod
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.4.0
    # via aiohttp
backoff==2.2.1
    # via runpod
backports-zstd==1.3.0
    # via aiohttp
bcrypt==5.0.0
    # via paramiko
boto3==1.42.40
    # via runpod
botocore==1.42.40
    # via
    #   boto3
    #   s3transfer
brotli==1.2.0
    # via aiohttp
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via
    #   cryptography
    #   pycares
    #   pynacl
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   rich-toolkit
    #   runpod
    #   typer
    #   uvicorn
cloudpickle==3.1.2
    # via runpod-flash
colorama==0.4.6
    # via runpod
cryptography==45.0.7
    # via
    #   paramiko
    #   runpod
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.128.0
    # via runpod
fastapi-cli==0.0.20
    # via fastapi
fastapi-cloud-cli==0.11.0
    # via fastapi-cli
fastar==0.8.0
    # via fastapi-cloud-cli
filelock==3.20.3
    # via runpod
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
inquirerpy==0.3.4
    # via runpod
invoke==2.2.1
    # via paramiko
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
numpy==2.4.2
    # via runpod-flash-examples (pyproject.toml)
orjson==3.11.7
    # via fastapi
paramiko==4.0.0
    # via runpod
pathspec==1.0.4
    # via runpod-flash
pfzy==0.3.4
    # via inquirerpy
pillow==12.1.0
    # via runpod-flash-examples (pyproject.toml)
prettytable==3.17.0
    # via runpod
prompt-toolkit==3.0.52
    # via
    #   inquirerpy
    #   questionary
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
py-cpuinfo==9.0.0
    # via runpod
pycares==5.0.1
    # via aiodns
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
    #   runpod-flash
pydantic-core==2.41.5
    # via pydantic
pydantic-extra-types==2.11.0
    # via fastapi
pydantic-settings==2.12.0
    # via fastapi
pygments==2.19.2
    # via rich
pynacl==1.6.2
    # via paramiko
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   runpod-flash
    #   uvicorn
python-multipart==0.0.22
    # via
    #   runpod-flash-examples (pyproject.toml)
    #   fastapi
pyyaml==6.0.3
    # via
    #   fastapi
    #   uvicorn
questionary==2.1.1
    # via runpod-flash
requests==2.32.5
    # via runpod
rich==14.3.2
    # via
    #   rich-toolkit
    #   runpod-flash
    #   typer
rich-toolkit==0.18.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
runpod==1.8.1
    # via runpod-flash
s3transfer==0.16.0
    # via boto3
sentry-sdk==2.51.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
starlette==0.50.0
    # via fastapi
structlog==25.5.0
    # via runpod-flash-examples (pyproject.toml)
runpod-flash==0.25.1
    # via runpod-flash-examples (pyproject.toml)
tomli==2.4.0
    # via runpod
tomlkit==0.14.0
    # via runpod
tqdm==4.67.2
    # via tqdm-loggable
tqdm-loggable==0.2
    # via runpod
typer==0.21.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
    #   runpod-flash
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.11.0
    # via fastapi
urllib3==2.6.3
    # via
    #   botocore
    #   requests
    #   runpod
    #   sentry-sdk
uvicorn==0.40.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.22.1
    # via uvicorn
watchdog==6.0.0
    # via runpod
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.5.3
    # via
    #   prettytable
    #   prompt-toolkit
websockets==16.0
    # via uvicorn
yarl==1.22.0
    # via aiohttp
